Defining Interfaces.
Note: We only create a business object class if that class provides some functionality that we want to use throughout our application - such as a calculateDiscount method for example.

If we want to add a business object at some point, then add the 'I' prefix to the interface and allow the business object to 'implements' the interface.

From https://angular.io/guide/styleguide#interfaces

- Consider naming an interface without an I prefix.
- Consider using a class instead of an interface.